+++ src/routes/Main/LeftContents/ButtonListComponent/LinkButton.tsx
@@ -1,4 +1,4 @@
-import Button from "../../../../components/common/Button";
+import Button from "../../../../components/common/SquareButton";
 
 type buttonProps = {
   icon: string;
